complete les service

+++ src/main/java/com/ouadia/rovista1/repositories/PaiementRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface PaiementRepository extends JpaRepository<Paiement, Long> {
+public interface PaiementRepository extends JpaRepository<Paiement, Integer> {
 
 }
